python/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
moviepy==1.0.3
pillow==10.4.0
pydub==0.25.1
google-cloud-translate==3.15.5
boto3==1.35.0
pydantic==2.9.2
srt==3.5.3
pytrends==4.9.2
serpapi==0.1.5
